refactor(main): extract run/config/connection/completion/output modules
Splits `src/main.rs` (~750 lines mixing CLI dispatch, config resolution,
TLS pool creation, query execution, output dispatch, completion
generation, and test harness) into focused modules:

- `src/main.rs` (~55 lines): parse CLI, init crypto + tracing, dispatch
  subcommands / `--dump-config`, hand off to `run::run(cli)`.
- `src/run.rs`: query-execution pipeline (resolve → connect → query →
  write) + MySQL-error-code mapper. Calls `config`, `connection`,
  `output`.
- `src/config.rs`: `resolve_database_url`, `resolve_database_query`,
  `resolve_output_file`, `dump_configuration`, `validate_query_file_path`
  and its size/extension constants.
- `src/connection.rs`: `create_database_connection` (TLS-aware pool
  builder).
- `src/completion.rs`: `generate_completion` for shell completions.
- `src/output.rs`: `write_output`, `create_output_file`, open-error
  classifier.

Unit tests moved alongside their subject (17 config, 3 connection, 4
completion = 24 tests). Each new module is published as `pub mod` under
`src/lib.rs` so future test code can reach them directly. No public-
binary behaviour changes: same CLI, same exit codes, same output
formats, same error messages.

Refs #014 (P1-K batch).
